+++ src/apis/auth.ts
@@ -1,19 +1,27 @@
-const rp = require('request-promise-native');
+const fetch = require('node-fetch');
 
 import { AuthReturnDataforRestaurantManagementGroupClient } from '../types/auth.types';
+import { ToastError } from '..';
 
-export async function requestNewToken(toastHostname: string, toastClient: string, toastSecret: string) : Promise<AuthReturnDataforRestaurantManagementGroupClient> {
+export async function requestNewToken(toastHostname: string, toastClient: string, toastSecret: string): Promise<AuthReturnDataforRestaurantManagementGroupClient> {
     const options = {
-        uri: `https://${toastHostname}/usermgmt/v1/oauth/token`,
         method: 'POST',
         body: `grant_type=client_credentials&client_id=${toastClient}&client_secret=${toastSecret}`,
         headers: {
             'Content-Type': 'application/x-www-form-urlencoded'
         },
-        json: true,
-    }  ;
+    };
 
-    const response = await rp(options);
+    try {
+        const response = await fetch(`https://${toastHostname}/usermgmt/v1/oauth/token`, options);
 
-    return response;
+        return response.json();
+    } catch (e) {
+        throw new ToastError({
+            message: e.message,
+            endpoint: 'request credentials',
+            path: '/usermgmt/v1/oauth/token'
+        });
+    }
+
 };