+++ src/index.ts
@@ -5,9 +5,13 @@ export * from './types/toast.types';
 export * from './types/restaurants.types'
 export * from './types/config.types'
 export * from './types/loyalty.types'
+export * from './types/labor.types';
+export * from './types/cash-management.types';
 
-export * from './apis/auth'
+export * from './apis/auth';
 export * from './apis/orders';
 export * from './apis/menus';
 export * from './apis/restaurants';
-export * from './apis/config'
+export * from './apis/config';
+export * from './apis/labor';
+export * from './apis/cash-management';