+++ src/types/config.types.ts
@@ -1,6 +1,6 @@
-import { ToastReference } from './toast.types';
+import { ToastReference, ExternalReference } from './toast.types';
 
-export class Discount extends ToastReference{
+export interface Discount extends ToastReference{
     name?: string;
     active?: boolean;
     type?: 'PERCENT' | 'FIXED' | 'OPEN_PERCENT' | 'OPEN_FIXED' | 'BOGO' | 'FIXED_TOTAL';
@@ -10,4 +10,46 @@ export class Discount extends ToastReference{
     nonExclusive?: boolean;
     itemPickingPriority?: 'FIRST' | 'LEAST_EXPENSIVE' | 'MOST_EXPENSIVE';
     fixedTotal?: number;
+}
+
+export interface RevenueCenter extends ToastReference {
+    name?: string;
+    description?: string;
+}
+
+export interface DiningOption extends ExternalReference {
+    name?: string;
+    behavior?: 'DINE_IN' | 'TAKE_OUT' | 'DELIVERY';
+}
+
+export interface CashDrawer extends ToastReference {
+    printer?: ToastReference;
+}
+
+export interface AlternatePaymentType extends ExternalReference {
+    name?: string;
+}
+
+export interface NoSaleReason extends ToastReference {
+    name?: string;
+}
+
+export interface PayoutReason extends ToastReference {
+    name?: string;
+}
+
+export interface VoidReason extends ToastReference {
+    name?: string;
+}
+
+export interface BreakType extends ToastReference {
+    name?: string;
+    active?: boolean;
+    paid?: boolean;
+    duration?: number;
+    enforceMinimumTime?: boolean;
+    trackMissedBreaks?: boolean;
+    breakIntervalHrs?: number;
+    breakIntervalMins?: number;
+    trackBreakAcknowledgement?: boolean;
 }